requirements.txt
Flask
psycopg2-binary
pandas
scikit-learn
joblib
python-dotenv
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Jinja2==3.1.2
